Relocate `PublishableFluentModel` into publishing app, re #22
- move `PublishableFluentModel` class from `articles` models to the
  central publishing app location that makes more sense
- relocate ICEKit mixins from `abstract_models` module to the new
  `mixins` module to avoid circular import problems
- fix import and document references to mixins to refer to new module.
