+++ icekit/articles/abstract_models.py
@@ -1,12 +1,6 @@
-from icekit.abstract_models import FluentFieldsMixin
-from icekit.publishing.models import PublishingModel
+from icekit.publishing.models import PublishableFluentModel
 from django.db import models
 
-# TODO: this should ideally be in icekit.abstract_models, but doing so
-# creates circular import errors.
-class PublishableFluentModel(FluentFieldsMixin, PublishingModel):
-    class Meta:
-        abstract = True
 
 class TitleSlugMixin(models.Model):
     # TODO: this should perhaps become part of a wider ICEkit mixin that covers
